+++ index/scatter_elements_v2/op_host/op_api/aclnn_scatter.cpp
@@ -114,7 +114,7 @@ static bool CompareTensorShape(const aclTensor* a, const aclTensor* b, int64_t d
     auto aDimSize = a->GetViewShape().GetDimNum();
     auto bDimSize = b->GetViewShape().GetDimNum();
     aDimSize = aDimSize < bDimSize ? aDimSize : bDimSize;
-    for (int64_t i = 0; i < aDimSize; i++) {
+    for (int64_t i = 0; i < static_cast<int64_t>(aDimSize); i++) {
         if (i != dim) {
             auto aDim = (a->GetViewShape())[i];
             auto bDim = (b->GetViewShape())[i];
@@ -428,7 +428,7 @@ static bool IsRouteToUpdate(const aclTensor *x, const op::Shape &selfShape, cons
         return false;
     }
 
-    for (size_t i = viewStrides.size() - 1; i >= 0; i--) {
+    for (int i = static_cast<int>(viewStrides.size()) - 1; i >= 0; i--) {
         if (viewStrides[i] != 0) {
             boardCastIdx = i;
             break;
@@ -542,7 +542,7 @@ static aclnnStatus ExecScatterBase(
 
     CHECK_COND(selfDimNum - 1 >= 0, ACLNN_ERR_PARAM_INVALID, "self dim num must greater than 0");
     std::vector<int64_t> perm(selfDimNum);
-    for (size_t i = 0; i < selfDimNum; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(selfDimNum); ++i) {
         perm[i] = i;
     }
     std::swap(perm[dimFinal], perm[selfDimNum - 1]);
